% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSlope.R
\name{generateSlope}
\alias{generateSlope}
\title{Slope Matrix}
\usage{
generateSlope(lfc, tp, group)
}
\arguments{
\item{lfc}{The numeric matrix or data.frame of log2FC data. Each row represents a gene and each column represents a sample.}

\item{tp}{A vector of time points of the samples in the matrix lfc. The length of the vector should be the same as the number of samples (i.e. the number of columns in the matrix \code{lfc}).}

\item{group}{A vector of indices indicating the set of samples from a particular drug/compound treatment. The (time-series) samples from the same drug treatment experiment should have the same unique index. The length of the vector should be the same as the number of samples.}
}
\value{
\item{slopeMat}{the slope matrix}
}
\description{
This function calculates a slope matrix from time-series log2FC expression data
}
\details{
The time-series samples from the same experiment should have the same group index. A repitition experiment should have a different group index. Together with the information of time points \code{tp}, this information is needed to calculate time slope matrix.

For more than two time points in an experiment, the \code{2nd order accuracy finite difference} is used for calculating the slopes of the corresponding samples. If only two time points are available in an experiment, the \code{linear slope} from the two time points is used for that experiment.
}
